# docs/swagger.yaml
basePath: /
definitions:
  controllers.AuthLoginStruct:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  enum.Role:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - Admin
    - User
...
  models.User:
    description: User represents a user with details such as name, email, password,
      etc.
    properties:
      address:
        description: |-
          Address of the user, optional
          @example "123 Main St, Springfield, IL"
        type: string
      age:
        description: |-
          Age of the user, optional
          @example 25
        type: integer
      created_at:
        description: |-
          Timestamp when the user was created
          @example "2025-04-10T14:22:00Z"
        type: string
      email:
        description: |-
          Email of the user, unique
          @example "johndoe@example.com"
        type: string
      gender:
        description: |-
          Gender of the user, optional
          @example "male"
        type: string
      id:
        description: |-
          ID of the user, automatically generated
          @example 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_verified:
        description: |-
          Whether the user's email has been verified
          @example false
        type: boolean
      name:
        description: |-
          Name of the user
          @example "John Doe"
        type: string
      password:
        description: |-
          Password of the user
          @example "password123"
        type: string
      phone:
        description: |-
          Phone number of the user, optional
          @example "+1234567890"
        type: string
      role:
        allOf:
        - $ref: '#/definitions/enum.Role'
        description: |-
          Role of the user, default is 'user'
          @example "admin"
      updated_at:
        description: |-
          Timestamp when the user was last updated
          @example "2025-04-10T14:22:00Z"
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: QFirst is a robust and easy-to-use booking system designed to help
    businesses manage appointments and reservations efficiently.
  title: QFirst API Documentation
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user by verifying their email and password. On successful
        authentication, a JWT token will be generated for the user.
      parameters:
      - description: User login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/controllers.AuthLoginStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input or missing data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in to the platform.
      tags:
      - Authentication
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Add a new user to the platform. The request must include valid
        user data, including name, email, and password. The user will be created and
        saved in the database.
      parameters:
      - description: User data to sign up
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input or missing data
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already in use
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Signing up to the platform.
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
  CsrfToken:
    description: CSRF protection token.
    in: header
    name: X-CSRF-Token
    type: apiKey
swagger: "2.0"